project selector: reorganize file (#1516)
